rework app navbar
